Tokenization will respect the order in which tokens are defined
From this commit, vst_tokenize() will arrange the tokens in the same
order as specified in the scheme file. This is required to improve
learning. With this change, learn should always learn most used tokens
before it gives up.

This commit also removes VST files. It needs to be compiled before using