[varnamc] varnamruby.rb will be loaded from the standard search path
Before this commit, varnamc does a require_relative and expected the
varnamruby.rb to be in the same path as varnamc. But varnamruby.rb is
not an executable but a library and should not be installd at bin.
Debian lintian was giving warning about this.

This commit adds a set of predefined search paths, ., /usr/local/lib and
/usr/lib. The first matched path will be taken into account