[build] Added VARNAM_LIB_ARCH environment variable
Setting this variable affects the way libraries are installed. If this
is set, install will create a subdirectory inside "lib" with the
variable value.

This is useful for installing libvarnam on environments where MultiArch
is supported. Debian and Ubuntu requires this. So the packaging rule
file will set this directory to whatever architecture the build is
happening.