[tests] Tests will use the local symbol files than the installed one
This is to fix build tools happy because tests was failing in debian
build where there won't anything installed.

This commit introduces a new env variable `VARNAM_SOURCES_ROOT`. Value
for this variable should point to the root of libvarnam's source
directory. Tests uses this variable and sets the load paths appropriately