Fix #166: Varnam outputs invalid combinations with chil letters in Malayalam
Varnam learnings has the word `kilivaathil => കിളിവാതിൽ`. When Varanm finds this word,
what it does is use the word plus tokenizes the rest of it. This gives chil combinations.
This PR adds a check for chil to replace the ending chil with its root consonant so that
proper grammatical combinations can happen.